refactor: createTableRow 별도 함수로 분리
